+++ gazebo/rendering/VisualPrivate.hh
@@ -76,7 +76,10 @@ namespace gazebo
                 visibilityFlags(GZ_VISIBILITY_ALL),
                 type(Visual::VT_ENTITY),
                 layer(0),
-                geomSize(ignition::math::Vector3d::One)
+                geomSize(ignition::math::Vector3d::One),
+                inheritTransparency(true),
+                wireframe(false)
+
       {
       }
 
@@ -201,6 +204,12 @@ namespace gazebo
 
       /// \brief Size of attached geometry
       public: ignition::math::Vector3d geomSize;
+
+      /// \brief True to inherit transparency from parent.
+      public: bool inheritTransparency;
+
+      /// \brief True if wireframe mode is enabled
+      public: bool wireframe;
     };
     /// \}
   }